the register a new user, login, token authentication and post are all in working order.
